requirements.txt
pyaudio==0.2.14
opencv-python==4.10.0.84
ffmpeg-python==0.2.0
numpy==2.1.0